backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
openai==1.6.1
python-dotenv==1.0.0
numpy==1.26.3

